Modified the code to be modular.
